refactor(core): Move Agent negotiation logic
