+++ agent/tests/unit/test_incoming_offer_handling.py
@@ -17,7 +17,8 @@ from app.schema.pydantic_models import (
     GPUModel,
     Region,
 )
-from app.policies.store import PolicyStore, negotiation_respond_to_make_offer
+from core.agent.app.policy.store import PolicyStore
+from app.policies.store import negotiation_respond_to_make_offer
 from app.utils.sqlite_client import SQLiteClient
 from app.utils.action_executor import _extract_initial_price_from_order
 from app.utils.validation import determine_strategy_from_order
@@ -556,7 +557,7 @@ class TestThreadCreationWithInitialPriceAndStrategy:
 
     @pytest.fixture
     def thread_store(self, temp_db):
-        from app.policies.negotiation_thread import NegotiationThreadStore
+        from core.agent.app.policy.negotiation_thread import NegotiationThreadStore
         sqlite_client = SQLiteClient(db_path=temp_db)
         return NegotiationThreadStore(sqlite_client=sqlite_client)
