+++ agent/tests/unit/test_price_negotiation.py
@@ -19,7 +19,7 @@ NOTE: Policy requires strategy to be specified. If no strategy, passes to next p
 
 import pytest
 from app.schema.pydantic_models import DecisionContext, NegotiationEvent
-from app.policies.store import PolicyStore
+from core.agent.app.policy.store import PolicyStore
 from app.utils.sqlite_client import SQLiteClient
 from core.agent.app.policy.evaluator import CallableEvaluator
 import tempfile
@@ -458,7 +458,7 @@ class TestMultipleBilateralNegotiations:
     @pytest.fixture
     def thread_store(self, temp_db):
         """Create a NegotiationThreadStore for multi-party tests."""
-        from app.policies.negotiation_thread import NegotiationThreadStore
+        from core.agent.app.policy.negotiation_thread import NegotiationThreadStore
         sqlite_client = SQLiteClient(db_path=temp_db)
         return NegotiationThreadStore(sqlite_client=sqlite_client)
